+++ packages/alias/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rollup/plugin-alias ChangeLog
 
+## v5.0.1
+
+_2023-10-05_
+
+### Bugfixes
+
+- fix: ensure rollup 4 compatibility [#1595](https://github.com/rollup/plugins/pull/1595)
+
 ## v5.0.0
 
 _2023-04-04_