+++ packages/commonjs/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @rollup/plugin-commonjs ChangeLog
 
+## v25.0.7
+
+_2023-10-15_
+
+### Bugfixes
+
+- fix: bump magic-string version [#1596](https://github.com/rollup/plugins/pull/1596)
+
+## v25.0.6
+
+_2023-10-15_
+
+### Bugfixes
+
+- fix: Keep the shebang at the top of the file content [#1610](https://github.com/rollup/plugins/pull/1610)
+
+## v25.0.5
+
+_2023-10-05_
+
+### Bugfixes
+
+- fix: ensure rollup 4 compatibility [#1595](https://github.com/rollup/plugins/pull/1595)
+
 ## v25.0.4
 
 _2023-08-11_