+++ packages/dynamic-import-vars/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @rollup/plugin-dynamic-import-vars ChangeLog
 
+## v2.0.7
+
+_2023-10-15_
+
+### Bugfixes
+
+- fix: bump magic-string version [#1596](https://github.com/rollup/plugins/pull/1596)
+
+## v2.0.6
+
+_2023-10-05_
+
+### Bugfixes
+
+- fix: ensure rollup 4 compatibility [#1595](https://github.com/rollup/plugins/pull/1595)
+
 ## v2.0.5
 
 _2023-08-11_