+++ packages/esm-shim/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @rollup/plugin-esm-shim ChangeLog
 
+## v0.1.4
+
+_2023-10-15_
+
+### Bugfixes
+
+- fix: bump magic-string version [#1596](https://github.com/rollup/plugins/pull/1596)
+
+## v1.0.3
+
+_2023-10-05_
+
+### Bugfixes
+
+- fix: ensure rollup 4 compatibility [#1595](https://github.com/rollup/plugins/pull/1595)
+
 ## v0.1.2
 
 _2023-08-26_