+++ packages/strip/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @rollup/plugin-strip ChangeLog
 
+## v3.0.4
+
+_2023-10-15_
+
+### Bugfixes
+
+- fix: bump magic-string version [#1596](https://github.com/rollup/plugins/pull/1596)
+
+## v3.0.3
+
+_2023-10-05_
+
+### Bugfixes
+
+- fix: ensure rollup 4 compatibility [#1595](https://github.com/rollup/plugins/pull/1595)
+
 ## v3.0.2
 
 _2022-12-17_