+++ packages/virtual/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rollup/plugin-virtual ChangeLog
 
+## v3.0.2
+
+_2023-10-05_
+
+### Bugfixes
+
+- fix: ensure rollup 4 compatibility [#1595](https://github.com/rollup/plugins/pull/1595)
+
 ## v3.0.1
 
 _2022-10-21_